feat: :sparkles: refactor API service structure and add admin, auth, manager, test, and user services
